리팩토링

+++ src/main/java/com/example/boardbvsj/service/ReplyService.java
@@ -2,16 +2,19 @@ package com.example.boardbvsj.service;
 
 import com.example.boardbvsj.dto.replyDto.ReplyRequestDto;
 import com.example.boardbvsj.dto.replyDto.ReplyResponseDto;
+import com.example.boardbvsj.dto.replyDto.ReplyUpdateDto;
 import com.example.boardbvsj.entity.Board;
 import com.example.boardbvsj.entity.Member;
 import com.example.boardbvsj.entity.Reply;
 import com.example.boardbvsj.exception.customException.BoardNotFoundException;
+import com.example.boardbvsj.exception.customException.DifferentUsernameException;
 import com.example.boardbvsj.exception.customException.ReplyNotFoundException;
 import com.example.boardbvsj.exception.customException.UserNotFoundException;
 import com.example.boardbvsj.repository.BoardRepository;
 import com.example.boardbvsj.repository.MemberRepository;
 import com.example.boardbvsj.repository.ReplyRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -27,7 +30,7 @@ public class ReplyService {
     private final MemberRepository memberRepository;
     private final BoardRepository boardRepository;
 
-    public List<ReplyResponseDto> findAll(Long boardId){
+    public List<ReplyResponseDto> findAll(Long boardId) {
         List<Reply> replies = replyRepository.findAllByBoard_Id(boardId);
         List<ReplyResponseDto> list = replies.stream()
                 .map(ReplyResponseDto::new)
@@ -36,35 +39,38 @@ public class ReplyService {
         return list;
     }
 
-    public ReplyResponseDto findOne(Long replyId){
-        return new ReplyResponseDto(replyRepository.findById(replyId)
-        .orElseThrow(ReplyNotFoundException::new));
-    }
-
     @Transactional
-    public void createReply(ReplyRequestDto replyRequestDto){
-        Board board=boardRepository.findById(replyRequestDto.getBoardId())
+    public void createReply(ReplyRequestDto replyRequestDto) {
+        Board board = boardRepository.findById(replyRequestDto.getBoardId())
                 .orElseThrow(BoardNotFoundException::new);
         Member member = memberRepository.findByUsername(replyRequestDto.getUsername())
                 .orElseThrow(UserNotFoundException::new);
         Reply reply = replyRequestDto.toEntity(replyRequestDto);
-        reply.setBoard(board);
-        reply.setMember(member);
+        reply.addBoard(board);
+        reply.addMember(member);
         replyRepository.save(reply);
     }
 
 
     @Transactional
-    public void deleteReply(Long replyId){
+    public void deleteReply(Long replyId) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
         Reply reply = replyRepository.findById(replyId)
                 .orElseThrow(ReplyNotFoundException::new);
+        if (!username.equals("admin") && !username.equals(reply.getMember().getUsername())) {
+            throw new DifferentUsernameException();
+        }
         replyRepository.delete(reply);
     }
 
     @Transactional
-    public void updateReply(Long replyId, String updateReplyContent){
-        Reply reply = replyRepository.findById(replyId)
+    public void updateReply(ReplyUpdateDto replyUpdateDto) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        Reply reply = replyRepository.findById(replyUpdateDto.getReplyId())
                 .orElseThrow(ReplyNotFoundException::new);
-        reply.changeText(updateReplyContent);
+        if (!username.equals(reply.getMember().getUsername())) {
+            throw new DifferentUsernameException();
+        }
+        reply.changeText(replyUpdateDto.getContent());
     }
 }
